+++ license_manager/apps/api/utils.py
@@ -1,18 +1,11 @@
 """ Utility functions. """
 import uuid
-from datetime import datetime
 
 from django.shortcuts import get_object_or_404
 from edx_rbac.utils import get_decoded_jwt
-from pytz import UTC
 from rest_framework.exceptions import ParseError
 
 from license_manager.apps.subscriptions.models import License, SubscriptionPlan
-
-
-def localized_utcnow():
-    """Helper function to return localized utcnow()."""
-    return UTC.localize(datetime.utcnow())  # pylint: disable=no-value-for-parameter
 
 
 def get_subscription_plan_from_enterprise(request):
@@ -80,15 +73,3 @@ def get_subscription_plan_by_activation_key(request):
         subscription_plan__is_active=True,
     )
     return user_license.subscription_plan
-
-
-def set_last_remind_date_to_now(licenses):
-    """
-    Helper function to bulk set the `last_remind_date` on a group of licenses to now.
-
-    Args:
-        licenses (iterable): The licenses to set the `last_remind_date` on.
-    """
-    for subscription_license in licenses:
-        subscription_license.last_remind_date = localized_utcnow()
-    License.objects.bulk_update(licenses, ['last_remind_date'])
