+++ license_manager/apps/subscriptions/utils.py
@@ -1,0 +1,9 @@
+""" Utility functions for the subscriptions app. """
+from datetime import datetime
+
+from pytz import UTC
+
+
+def localized_utcnow():
+    """Helper function to return localized utcnow()."""
+    return UTC.localize(datetime.utcnow())  # pylint: disable=no-value-for-parameter
